requirements-tx.txt
qrcode[pil]==8.2
pillow==12.1.1